+++ src/package-tests/cjs/cjs.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import Console from 'react-component-console';
+import Console from '@haensl/react-component-console';
 
 describe('esm module test', () => {
   let container;